# Optimize looking up min/max values of pages with dictionaries