--- a/dictionary_default.go
+++ b/dictionary_default.go
@@ -27,19 +27,23 @@ func (d *booleanDictionary) Type() Type { return newIndexedType(d.typ, d) }
 
 func (d *booleanDictionary) Len() int { return 2 }
 
-func (d *booleanDictionary) Index(i int) Value { return makeValueBoolean(d.values[i]) }
+func (d *booleanDictionary) Index(i int32) Value { return makeValueBoolean(d.values[i]) }
 
-func (d *booleanDictionary) Insert(v Value) int {
-	if v.Boolean() {
-		return 1
-	} else {
-		return 0
+func (d *booleanDictionary) Insert(indexes []int32, values []Value) {
+	_ = indexes[:len(values)]
+
+	for i, v := range values {
+		if v.Boolean() {
+			indexes[i] = 1
+		} else {
+			indexes[i] = 0
+		}
 	}
 }
 
 func (d *booleanDictionary) Lookup(indexes []int32, values []Value) {
 	for i, j := range indexes {
-		values[i] = d.Index(int(j))
+		values[i] = d.Index(j)
 	}
 }
 
@@ -105,29 +109,35 @@ func (d *int32Dictionary) Type() Type { return newIndexedType(d.typ, d) }
 
 func (d *int32Dictionary) Len() int { return len(d.values) }
 
-func (d *int32Dictionary) Index(i int) Value { return makeValueInt32(d.values[i]) }
+func (d *int32Dictionary) Index(i int32) Value { return makeValueInt32(d.values[i]) }
 
-func (d *int32Dictionary) Insert(v Value) int { return d.insert(v.Int32()) }
+func (d *int32Dictionary) Insert(indexes []int32, values []Value) {
+	_ = indexes[:len(values)]
 
-func (d *int32Dictionary) insert(value int32) int {
-	if index, exists := d.index[value]; exists {
-		return int(index)
-	}
 	if d.index == nil {
 		d.index = make(map[int32]int32, cap(d.values))
 		for i, v := range d.values {
 			d.index[v] = int32(i)
 		}
 	}
-	index := len(d.values)
-	d.index[value] = int32(index)
-	d.values = append(d.values, value)
-	return index
+
+	for i, v := range values {
+		value := v.Int32()
+
+		index, exists := d.index[value]
+		if !exists {
+			index = int32(len(d.values))
+			d.values = append(d.values, value)
+			d.index[value] = index
+		}
+
+		indexes[i] = index
+	}
 }
 
 func (d *int32Dictionary) Lookup(indexes []int32, values []Value) {
 	for i, j := range indexes {
-		values[i] = d.Index(int(j))
+		values[i] = d.Index(j)
 	}
 }
 
@@ -206,29 +216,35 @@ func (d *int64Dictionary) Type() Type { return newIndexedType(d.typ, d) }
 
 func (d *int64Dictionary) Len() int { return len(d.values) }
 
-func (d *int64Dictionary) Index(i int) Value { return makeValueInt64(d.values[i]) }
+func (d *int64Dictionary) Index(i int32) Value { return makeValueInt64(d.values[i]) }
 
-func (d *int64Dictionary) Insert(v Value) int { return d.insert(v.Int64()) }
+func (d *int64Dictionary) Insert(indexes []int32, values []Value) {
+	_ = indexes[:len(values)]
 
-func (d *int64Dictionary) insert(value int64) int {
-	if index, exists := d.index[value]; exists {
-		return int(index)
-	}
 	if d.index == nil {
 		d.index = make(map[int64]int32, cap(d.values))
 		for i, v := range d.values {
 			d.index[v] = int32(i)
 		}
 	}
-	index := len(d.values)
-	d.index[value] = int32(index)
-	d.values = append(d.values, value)
-	return index
+
+	for i, v := range values {
+		value := v.Int64()
+
+		index, exists := d.index[value]
+		if !exists {
+			index = int32(len(d.values))
+			d.values = append(d.values, value)
+			d.index[value] = index
+		}
+
+		indexes[i] = index
+	}
 }
 
 func (d *int64Dictionary) Lookup(indexes []int32, values []Value) {
 	for i, j := range indexes {
-		values[i] = d.Index(int(j))
+		values[i] = d.Index(j)
 	}
 }
 
@@ -307,29 +323,35 @@ func (d *int96Dictionary) Type() Type { return newIndexedType(d.typ, d) }
 
 func (d *int96Dictionary) Len() int { return len(d.values) }
 
-func (d *int96Dictionary) Index(i int) Value { return makeValueInt96(d.values[i]) }
+func (d *int96Dictionary) Index(i int32) Value { return makeValueInt96(d.values[i]) }
 
-func (d *int96Dictionary) Insert(v Value) int { return d.insert(v.Int96()) }
+func (d *int96Dictionary) Insert(indexes []int32, values []Value) {
+	_ = indexes[:len(values)]
 
-func (d *int96Dictionary) insert(value deprecated.Int96) int {
-	if index, exists := d.index[value]; exists {
-		return int(index)
-	}
 	if d.index == nil {
 		d.index = make(map[deprecated.Int96]int32, cap(d.values))
 		for i, v := range d.values {
 			d.index[v] = int32(i)
 		}
 	}
-	index := len(d.values)
-	d.index[value] = int32(index)
-	d.values = append(d.values, value)
-	return index
+
+	for i, v := range values {
+		value := v.Int96()
+
+		index, exists := d.index[value]
+		if !exists {
+			index = int32(len(d.values))
+			d.values = append(d.values, value)
+			d.index[value] = index
+		}
+
+		indexes[i] = index
+	}
 }
 
 func (d *int96Dictionary) Lookup(indexes []int32, values []Value) {
 	for i, j := range indexes {
-		values[i] = d.Index(int(j))
+		values[i] = d.Index(j)
 	}
 }
 
@@ -408,29 +430,35 @@ func (d *floatDictionary) Type() Type { return newIndexedType(d.typ, d) }
 
 func (d *floatDictionary) Len() int { return len(d.values) }
 
-func (d *floatDictionary) Index(i int) Value { return makeValueFloat(d.values[i]) }
+func (d *floatDictionary) Index(i int32) Value { return makeValueFloat(d.values[i]) }
 
-func (d *floatDictionary) Insert(v Value) int { return d.insert(v.Float()) }
+func (d *floatDictionary) Insert(indexes []int32, values []Value) {
+	_ = indexes[:len(values)]
 
-func (d *floatDictionary) insert(value float32) int {
-	if index, exists := d.index[value]; exists {
-		return int(index)
-	}
 	if d.index == nil {
 		d.index = make(map[float32]int32, cap(d.values))
 		for i, v := range d.values {
 			d.index[v] = int32(i)
 		}
 	}
-	index := len(d.values)
-	d.index[value] = int32(index)
-	d.values = append(d.values, value)
-	return index
+
+	for i, v := range values {
+		value := v.Float()
+
+		index, exists := d.index[value]
+		if !exists {
+			index = int32(len(d.values))
+			d.values = append(d.values, value)
+			d.index[value] = index
+		}
+
+		indexes[i] = index
+	}
 }
 
 func (d *floatDictionary) Lookup(indexes []int32, values []Value) {
 	for i, j := range indexes {
-		values[i] = d.Index(int(j))
+		values[i] = d.Index(j)
 	}
 }
 
@@ -509,29 +537,35 @@ func (d *doubleDictionary) Type() Type { return newIndexedType(d.typ, d) }
 
 func (d *doubleDictionary) Len() int { return len(d.values) }
 
-func (d *doubleDictionary) Index(i int) Value { return makeValueDouble(d.values[i]) }
+func (d *doubleDictionary) Index(i int32) Value { return makeValueDouble(d.values[i]) }
 
-func (d *doubleDictionary) Insert(v Value) int { return d.insert(v.Double()) }
+func (d *doubleDictionary) Insert(indexes []int32, values []Value) {
+	_ = indexes[:len(values)]
 
-func (d *doubleDictionary) insert(value float64) int {
-	if index, exists := d.index[value]; exists {
-		return int(index)
-	}
 	if d.index == nil {
 		d.index = make(map[float64]int32, cap(d.values))
 		for i, v := range d.values {
 			d.index[v] = int32(i)
 		}
 	}
-	index := len(d.values)
-	d.index[value] = int32(index)
-	d.values = append(d.values, value)
-	return index
+
+	for i, v := range values {
+		value := v.Double()
+
+		index, exists := d.index[value]
+		if !exists {
+			index = int32(len(d.values))
+			d.values = append(d.values, value)
+			d.index[value] = index
+		}
+
+		indexes[i] = index
+	}
 }
 
 func (d *doubleDictionary) Lookup(indexes []int32, values []Value) {
 	for i, j := range indexes {
-		values[i] = d.Index(int(j))
+		values[i] = d.Index(j)
 	}
 }