This PR adds a benchmark to highlight a performance bottleneck that was reported when computing the min/max bounds of pages with dictionaries, and proposes introducing a new `parquet.Dictionary.Bounds` method optimizing the operation.

After analyzing CPU profiles, it appeared that the excessive compute footprint was mostly driven by the cost of abstractions: multiple indirect method calls through interface and deep call stacks in a tight loop over large data sets add up quickly.

To amortize the cost of abstractions, I am adding a `Bounds(indexes []int32) (min, max Value)` method to the `parquet.Dictionary` interface, that the indexed pages delegate the computation of min/max values to, passing the list of dictionary indexes that belong to the page as argument. With this model, we cross abstraction boundaries only once per bound lookup, instead of once for each comparison, which amortizes the compute cost.

Here is a comparison of the benchmark I added highlighting the impact of the change:
```
name                             old time/op    new time/op    delta
DictionaryPageBounds/BOOLEAN       58.9µs ± 1%    27.6µs ± 1%  -53.22%  (p=0.000 n=10+9)
DictionaryPageBounds/INT32         65.6µs ± 0%    38.4µs ± 1%  -41.49%  (p=0.000 n=10+10)
DictionaryPageBounds/INT64         69.9µs ± 1%    42.0µs ± 1%  -39.89%  (p=0.000 n=9+10)
DictionaryPageBounds/INT96          127µs ± 1%      64µs ± 1%  -49.41%  (p=0.000 n=8+10)
DictionaryPageBounds/FLOAT         73.0µs ± 1%    46.0µs ± 1%  -37.02%  (p=0.000 n=10+10)
DictionaryPageBounds/DOUBLE        74.9µs ± 0%    47.7µs ± 0%  -36.38%  (p=0.000 n=10+6)
DictionaryPageBounds/BYTE_ARRAY    70.1µs ± 1%    41.0µs ± 1%  -41.51%  (p=0.000 n=10+10)

name                             old alloc/op   new alloc/op   delta
DictionaryPageBounds/BOOLEAN        331kB ± 0%     331kB ± 0%   -0.01%  (p=0.000 n=10+10)
DictionaryPageBounds/INT32          303kB ± 0%     303kB ± 0%   -0.01%  (p=0.000 n=10+10)
DictionaryPageBounds/INT64          393kB ± 0%     393kB ± 0%   -0.01%  (p=0.000 n=10+10)
DictionaryPageBounds/INT96          491kB ± 0%     475kB ± 0%   -3.26%  (p=0.000 n=10+10)
DictionaryPageBounds/FLOAT          303kB ± 0%     303kB ± 0%   -0.01%  (p=0.000 n=10+10)
DictionaryPageBounds/DOUBLE         393kB ± 0%     393kB ± 0%   -0.01%  (p=0.000 n=10+10)
DictionaryPageBounds/BYTE_ARRAY     361kB ± 0%     361kB ± 0%   -0.01%  (p=0.000 n=10+10)

name                             old allocs/op  new allocs/op  delta
DictionaryPageBounds/BOOLEAN         9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.000 n=10+10)
DictionaryPageBounds/INT32           9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.000 n=10+10)
DictionaryPageBounds/INT64           9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.000 n=10+10)
DictionaryPageBounds/INT96          1.01k ± 0%     0.01k ± 0%  -99.01%  (p=0.000 n=10+10)
DictionaryPageBounds/FLOAT           9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.000 n=10+10)
DictionaryPageBounds/DOUBLE          9.00 ± 0%      8.00 ± 0%  -11.11%  (p=0.000 n=10+10)
DictionaryPageBounds/BYTE_ARRAY      10.0 ± 0%       9.0 ± 0%  -10.00%  (p=0.000 n=10+10)
```

Another interesting quality of this approach is that it opens-up the door to further optimizations using SIMD operations and AVX gather instructions since we are working on two arrays of contiguous values (the indexes and the dictionary values).

I based the PR on #103 since I had made changes to the dictionary types. I gotta say I'm looking forward to not have to maintain 1.17 anymore :D